Refactor UI test contexts out a little.
